Extract generic Picker component
